modify the handedness chain
modify the gltf_loader so it convert (RH + CCW) data to (LH + CW). The camera and projection will always using LH system due to the shader part. Note: the blender exported gltf would has fliped normal due to the winding order, which is incorrect as gltf. So now just force flip it.