implement oct-tree
1. Initial implement of build oct-tree.
The compute shader pass start with semi-side-length of the entire voxel scene. If it was the first level, then the child[ii] stores the index of the texel buffer calculated in the voxlization pass. In the other condition, it stores the index of child node in the oct-tree.
The question is: should I store [the level], [the side-length], [the center pos] of the oct-tree node, or I can just calculate it in the following pass.