backup
Fix some set binding issue. Now have to take a look about the VK_EXT_graphics_pipeline_library. The new implement introduced a nullptr as a set layout since the bindless reserved for set==1 and could be null. This is not allowed if VK_EXT_graphics_pipeline_library not enable